changed station responses to fit the new scheme
