+++ src/api/sitzung.rs
@@ -562,7 +562,11 @@ pub async fn kal_get_by_param(
         hparams.if_modified_since,
     );
     if result.is_none() {
-        return Ok(KalGetResponse::Status416_RequestRangeNotSatisfiable);
+        return Ok(KalGetResponse::Status416_RequestRangeNotSatisfiable {
+            x_rate_limit_limit: None,
+            x_rate_limit_remaining: None,
+            x_rate_limit_reset: None,
+        });
     }
 
     let params = retrieve::SitzungFilterParameters {
@@ -579,7 +583,11 @@ pub async fn kal_get_by_param(
     // retrieval
     let result = retrieve::sitzung_by_param(&params, tx).await?;
     if result.is_empty() {
-        Ok(KalGetResponse::Status204_NoContentFoundForTheSpecifiedParameters)
+        Ok(KalGetResponse::Status204_NoContent {
+            x_rate_limit_limit: None,
+            x_rate_limit_remaining: None,
+            x_rate_limit_reset: None,
+        })
     } else {
         Ok(KalGetResponse::Status200_AntwortAufEineGefilterteAnfrageZuSitzungen(result))
     }
@@ -718,5 +726,829 @@ mod test_applicable_date_range {
         let result = result.unwrap();
         assert!(result.since.is_some() && result.since.unwrap() == expected_since);
         assert!(result.until.is_some() && result.until.unwrap() == expected_until);
+    }
+}
+
+#[cfg(test)]
+mod sitzung_test {
+    use axum::http::Method;
+    use axum_extra::extract::{CookieJar, Host};
+    use chrono::Utc;
+    use openapi::apis::adminschnittstellen_collector_schnittstellen_kalender_sitzungen::*;
+    use openapi::apis::adminschnittstellen_sitzungen::*;
+    use openapi::apis::kalender_sitzungen_unauthorisiert::*;
+    use openapi::apis::sitzungen_unauthorisiert::*;
+
+    use openapi::models;
+    use uuid::Uuid;
+
+    use super::super::auth;
+    use super::super::endpoint_test::*;
+
+    // Calendar tests
+    #[tokio::test]
+    async fn test_calendar_endpoints() {
+        // Setup test server and database
+        let server = setup_server("test_calendar").await.unwrap();
+        let host = Host("localhost".to_string());
+        let cookies = CookieJar::new();
+
+        // Create test calendar entry
+        let test_date = chrono::Utc::now().date_naive();
+        let recent_date = test_date; // Define recent_date at the same scope level
+        let test_session = create_test_session();
+        let test_sessions = vec![test_session.clone()];
+
+        // Test cases for kal_date_put:
+        // 1. Update calendar entry with valid data and proper permissions
+        {
+            let response = server
+                .kal_date_put(
+                    &Method::PUT,
+                    &host,
+                    &cookies,
+                    &(auth::APIScope::Admin, 1),
+                    &models::KalDatePutHeaderParams {
+                        x_scraper_id: Uuid::nil(),
+                    },
+                    &models::KalDatePutPathParams {
+                        datum: test_date,
+                        parlament: models::Parlament::Bt,
+                    },
+                    &test_sessions,
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                KalDatePutResponse::Status201_Created {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+            // Allow time for database operations to complete
+            tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+        }
+
+        // 2. Update calendar entry with insufficient permissions
+        {
+            let response = server
+                .kal_date_put(
+                    &Method::PUT,
+                    &host,
+                    &cookies,
+                    &(auth::APIScope::Collector, 1), // Using Collector scope with old date should fail
+                    &models::KalDatePutHeaderParams {
+                        x_scraper_id: Uuid::nil(),
+                    },
+                    &models::KalDatePutPathParams {
+                        datum: test_date.checked_sub_days(chrono::Days::new(5)).unwrap(), // Date more than 1 day old
+                        parlament: models::Parlament::Bt,
+                    },
+                    &test_sessions,
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                KalDatePutResponse::Status403_AuthenticationFailed {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+
+        // 3. Update calendar entry with date constraints (collector is allowed to update recent dates)
+        {
+            // Use the already defined recent_date variable instead of redefining it
+            let response = server
+                .kal_date_put(
+                    &Method::PUT,
+                    &host,
+                    &cookies,
+                    &(auth::APIScope::Collector, 1),
+                    &models::KalDatePutHeaderParams {
+                        x_scraper_id: Uuid::nil(),
+                    },
+                    &models::KalDatePutPathParams {
+                        datum: recent_date,
+                        parlament: models::Parlament::Bt,
+                    },
+                    &test_sessions,
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                KalDatePutResponse::Status201_Created {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+            // Allow time for database operations to complete
+            tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+        }
+
+        // Test cases for kal_date_get:
+        // 1. Get calendar entry for valid date and parliament
+        {
+            let response = server
+                .kal_date_get(
+                    &Method::GET,
+                    &host,
+                    &cookies,
+                    &models::KalDateGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::KalDateGetPathParams {
+                        datum: recent_date,
+                        parlament: models::Parlament::Bt,
+                    },
+                    &models::KalDateGetQueryParams {
+                        page: None,
+                        per_page: None,
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                KalDateGetResponse::Status200_SuccessfulResponse { body, .. } => {
+                    assert!(!body.is_empty(), "Expected to find at least one session");
+                    assert_eq!(body[0].gremium.parlament, models::Parlament::Bt);
+                    assert_eq!(
+                        body[0].termin.date_naive(),
+                        recent_date,
+                        "Expected to find a session with the requested date"
+                    );
+                }
+                _ => panic!("Expected to find sessions for the valid date"),
+            }
+        }
+
+        // 2. Get calendar entry for non-existent date
+        {
+            let non_existent_date = chrono::NaiveDate::from_ymd_opt(1900, 1, 1).unwrap();
+            let response = server
+                .kal_date_get(
+                    &Method::GET,
+                    &host,
+                    &cookies,
+                    &models::KalDateGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::KalDateGetPathParams {
+                        datum: non_existent_date,
+                        parlament: models::Parlament::Bt,
+                    },
+                    &models::KalDateGetQueryParams {
+                        page: None,
+                        per_page: None,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                KalDateGetResponse::Status404_NotFound {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+        // TODO: Test for Status304_NotModified with set If-Modified-Since Header
+
+        // Test cases for kal_get:
+        // 1. Get calendar entries with valid parameters
+        {
+            let response = server
+                .kal_get(
+                    &Method::GET,
+                    &host,
+                    &cookies,
+                    &models::KalGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::KalGetQueryParams {
+                        y: Some(recent_date.format("%Y").to_string().parse::<i32>().unwrap()),
+                        m: Some(recent_date.format("%m").to_string().parse::<i32>().unwrap()),
+                        dom: None,
+                        gr: None,
+                        limit: Some(10),
+                        offset: Some(0),
+                        p: Some(models::Parlament::Bt),
+                        since: None,
+                        until: None,
+                        wp: Some(20),
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                KalGetResponse::Status200_SuccessfulResponse { body, .. } => {
+                    assert!(
+                        !body.is_empty(),
+                        "Expected to find sessions with valid filters"
+                    );
+                }
+                _ => panic!("Expected to find sessions with valid filters"),
+            }
+        }
+
+        // 2. Get calendar entries with invalid parameters (since > until)
+        {
+            let response = server
+                .kal_get(
+                    &Method::GET,
+                    &host,
+                    &cookies,
+                    &models::KalGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::KalGetQueryParams {
+                        y: None,
+                        m: None,
+                        dom: None,
+                        gr: None,
+                        limit: None,
+                        offset: None,
+                        p: None,
+                        since: Some(chrono::Utc::now()),
+                        until: Some(chrono::Utc::now() - chrono::Duration::days(1)), // until is before since
+                        wp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                KalGetResponse::Status416_RequestRangeNotSatisfiable {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+
+        // 3. Get calendar entries with date range
+        {
+            let start_date = recent_date
+                .and_time(chrono::NaiveTime::from_hms_opt(0, 0, 0).unwrap())
+                .and_utc();
+            let end_date = recent_date
+                .checked_add_days(chrono::Days::new(5))
+                .unwrap()
+                .and_time(chrono::NaiveTime::from_hms_opt(23, 59, 59).unwrap())
+                .and_utc();
+
+            let response = server
+                .kal_get(
+                    &Method::GET,
+                    &host,
+                    &cookies,
+                    &models::KalGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::KalGetQueryParams {
+                        y: None,
+                        m: None,
+                        dom: None,
+                        gr: None,
+                        limit: None,
+                        offset: None,
+                        p: Some(models::Parlament::Bt),
+                        since: Some(start_date),
+                        until: Some(end_date),
+                        wp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                KalGetResponse::Status200_SuccessfulResponse { body, .. } => {
+                    assert!(!body.is_empty(), "Expected to find sessions in date range");
+                    for session in body {
+                        assert!(
+                            session.termin >= start_date && session.termin <= end_date,
+                            "Found session outside requested date range"
+                        );
+                    }
+                }
+                _ => panic!("Expected to find sessions in date range"),
+            }
+        }
+
+        // TODO: Test for Status304_NotModified with set If-Modified-Since Header
+        // TODO: Test for Status204_NoContent
+
+        // Cleanup
+        cleanup_server("test_calendar").await.unwrap();
+    }
+
+    #[tokio::test]
+    pub(crate) async fn test_session_get_endpoints() {
+        // Setup test server and database
+        let server = setup_server("test_session_get").await.unwrap();
+
+        let test_session = create_test_session();
+        // First create the session
+        let create_response = server
+            .sid_put(
+                &Method::PUT,
+                &Host("localhost".to_string()),
+                &CookieJar::new(),
+                &(auth::APIScope::Admin, 1),
+                &models::SidPutPathParams {
+                    sid: test_session.api_id.unwrap(),
+                },
+                &test_session,
+            )
+            .await
+            .unwrap();
+        assert_eq!(
+            create_response,
+            SidPutResponse::Status201_Created {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None
+            }
+        );
+        tokio::time::sleep(tokio::time::Duration::from_secs(10)).await;
+        // Test cases for s_get_by_id:
+        // 1. Get existing session
+        {
+            let response = server
+                .s_get_by_id(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetByIdHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetByIdPathParams {
+                        sid: test_session.api_id.unwrap(),
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                SGetByIdResponse::Status200_Success { body, .. } => {
+                    assert_eq!(body.api_id, test_session.api_id);
+                    assert_eq!(body.titel, test_session.titel);
+                }
+                _ => panic!("Expected successful operation response"),
+            }
+        }
+
+        // 2. Get non-existent session
+        {
+            let non_existent_id = Uuid::now_v7();
+            let response = server
+                .s_get_by_id(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetByIdHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetByIdPathParams {
+                        sid: non_existent_id,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetByIdResponse::Status404_NotFound {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+
+        // 3. Get session with invalid ID
+        {
+            let invalid_id = Uuid::nil();
+            let response = server
+                .s_get_by_id(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetByIdHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetByIdPathParams { sid: invalid_id },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetByIdResponse::Status404_NotFound {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+        {
+            let response = server
+                .s_get_by_id(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetByIdHeaderParams {
+                        if_modified_since: Some(chrono::Utc::now()),
+                    },
+                    &models::SGetByIdPathParams {
+                        sid: test_session.api_id.unwrap(),
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetByIdResponse::Status304_NotModified {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+
+        // Test cases for s_get:
+        // 1. Get sessions with valid parameters
+        {
+            let response = server
+                .s_get(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetQueryParams {
+                        limit: None,
+                        offset: None,
+                        p: Some(models::Parlament::Bt),
+                        since: None,
+                        until: None,
+                        wp: Some(20),
+                        vgid: None,
+                        vgtyp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                SGetResponse::Status200_SuccessfulResponse { body, .. } => {
+                    assert!(!body.is_empty());
+                }
+                rsp => panic!("Expected successful operation response, got {:?}", rsp),
+            }
+        }
+
+        // 2. Get sessions with invalid parameters
+        {
+            let response = server
+                .s_get(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetQueryParams {
+                        limit: None,
+                        offset: None,
+                        p: None,
+                        since: Some(Utc::now()),
+                        until: Some(Utc::now() - chrono::Duration::days(365)),
+                        wp: None,
+                        vgid: None,
+                        vgtyp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetResponse::Status416_RequestRangeNotSatisfiable {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+
+        let test_session = create_test_session();
+        // First create a session with specific parameters
+        let create_response = server
+            .sid_put(
+                &Method::PUT,
+                &Host("localhost".to_string()),
+                &CookieJar::new(),
+                &(auth::APIScope::Admin, 1),
+                &models::SidPutPathParams {
+                    sid: test_session.api_id.unwrap(),
+                },
+                &test_session,
+            )
+            .await
+            .unwrap();
+        assert_eq!(
+            create_response,
+            SidPutResponse::Status201_Created {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None
+            }
+        );
+        tokio::time::sleep(tokio::time::Duration::from_secs(10)).await;
+        // 3. Get sessions with filters
+        {
+            let response = server
+                .s_get(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetQueryParams {
+                        limit: Some(10),
+                        offset: Some(0),
+                        p: Some(models::Parlament::Bt),
+                        since: None,
+                        until: None,
+                        wp: Some(20),
+                        vgid: None,
+                        vgtyp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            match response {
+                SGetResponse::Status200_SuccessfulResponse { body, .. } => {
+                    assert!(!body.is_empty());
+                }
+                _ => panic!("Expected successful operation response"),
+            }
+        }
+        {
+            let response = server
+                .s_get(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetHeaderParams {
+                        if_modified_since: Some(chrono::Utc::now()),
+                    },
+                    &models::SGetQueryParams {
+                        limit: Some(10),
+                        offset: Some(0),
+                        p: Some(models::Parlament::Bt),
+                        since: None,
+                        until: None,
+                        wp: Some(20),
+                        vgid: None,
+                        vgtyp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetResponse::Status304_NotModified {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+        {
+            let response = server
+                .s_get(
+                    &Method::GET,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetHeaderParams {
+                        if_modified_since: None,
+                    },
+                    &models::SGetQueryParams {
+                        limit: Some(10),
+                        offset: Some(0),
+                        p: Some(models::Parlament::Bt),
+                        since: None,
+                        until: None,
+                        wp: Some(22),
+                        vgid: None,
+                        vgtyp: None,
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SGetResponse::Status204_NoContent {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+        // Cleanup
+        cleanup_server("test_session_get").await.unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_session_modify_endpoints() {
+        let server = setup_server("session_modify_ep").await.unwrap();
+        let sitzung = create_test_session();
+        // - Input non-existing session
+        {
+            let response = server
+                .sid_put(
+                    &Method::PUT,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &(auth::APIScope::Admin, 1),
+                    &models::SidPutPathParams {
+                        sid: sitzung.api_id.unwrap(),
+                    },
+                    &sitzung,
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SidPutResponse::Status201_Created {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+            tokio::time::sleep(tokio::time::Duration::from_secs(10)).await;
+        }
+
+        // - Update existing session with the same data
+        let response = server
+            .sid_put(
+                &Method::PUT,
+                &Host("localhost".to_string()),
+                &CookieJar::new(),
+                &(auth::APIScope::Admin, 1),
+                &models::SidPutPathParams {
+                    sid: sitzung.api_id.unwrap(),
+                },
+                &sitzung,
+            )
+            .await
+            .unwrap();
+        assert_eq!(
+            response,
+            SidPutResponse::Status304_NotModified {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None
+            },
+            "Failed to update existing session with the same data.\nInput:  {:?}\n\n Output: {:?}",
+            sitzung,
+            server
+                .s_get_by_id(
+                    &Method::PUT,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &models::SGetByIdHeaderParams {
+                        if_modified_since: None
+                    },
+                    &models::SGetByIdPathParams {
+                        sid: sitzung.api_id.unwrap()
+                    },
+                )
+                .await
+                .unwrap()
+        );
+        tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+
+        // - Update existing session with valid new data
+        let rsp_new = models::Sitzung {
+            link: Some("https://example.com/a/b/c".to_string()),
+            ..sitzung.clone()
+        };
+        let response = server
+            .sid_put(
+                &Method::PUT,
+                &Host("localhost".to_string()),
+                &CookieJar::new(),
+                &(auth::APIScope::Admin, 1),
+                &models::SidPutPathParams {
+                    sid: sitzung.api_id.unwrap(),
+                },
+                &rsp_new,
+            )
+            .await
+            .unwrap();
+        assert_eq!(
+            response,
+            SidPutResponse::Status201_Created {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None
+            }
+        );
+        tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+
+        // - Update session with insufficient permissions
+        let response = server
+            .sid_put(
+                &Method::PUT,
+                &Host("localhost".to_string()),
+                &CookieJar::new(),
+                &(auth::APIScope::Collector, 1),
+                &models::SidPutPathParams {
+                    sid: sitzung.api_id.unwrap(),
+                },
+                &sitzung,
+            )
+            .await
+            .unwrap();
+        assert_eq!(
+            response,
+            SidPutResponse::Status403_AuthenticationFailed {
+                x_rate_limit_limit: None,
+                x_rate_limit_remaining: None,
+                x_rate_limit_reset: None
+            }
+        );
+
+        // Test cases for sitzung_delete:
+        {
+            // - Delete existing session with proper permissions
+            let response = server
+                .sitzung_delete(
+                    &Method::PUT,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &(auth::APIScope::KeyAdder, 1),
+                    &models::SitzungDeletePathParams {
+                        sid: sitzung.api_id.unwrap(),
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SitzungDeleteResponse::Status204_NoContent {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+
+            // - Delete non-existent session
+            let response = server
+                .sitzung_delete(
+                    &Method::PUT,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &(auth::APIScope::KeyAdder, 1),
+                    &models::SitzungDeletePathParams {
+                        sid: sitzung.api_id.unwrap(),
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SitzungDeleteResponse::Status404_NoElementWithThisID
+            );
+
+            // - Delete session with insufficient permissions
+            let response = server
+                .sitzung_delete(
+                    &Method::PUT,
+                    &Host("localhost".to_string()),
+                    &CookieJar::new(),
+                    &(auth::APIScope::Collector, 1),
+                    &models::SitzungDeletePathParams {
+                        sid: sitzung.api_id.unwrap(),
+                    },
+                )
+                .await
+                .unwrap();
+            assert_eq!(
+                response,
+                SitzungDeleteResponse::Status403_AuthenticationFailed {
+                    x_rate_limit_limit: None,
+                    x_rate_limit_remaining: None,
+                    x_rate_limit_reset: None
+                }
+            );
+        }
+        cleanup_server("session_modify_ep").await.unwrap();
     }
 }
