+++ src/main.rs
@@ -17,16 +17,6 @@ use tower_governor::{governor::GovernorConfigBuilder, *};
 
 pub use api::{LTZFArc, LTZFServer};
 pub use error::Result;
-use openapi::apis::{
-    adminschnittstellen_autoren::*,
-    adminschnittstellen_collector_schnittstellen_kalender_sitzungen::*,
-    adminschnittstellen_dokumente::*, adminschnittstellen_enumerations::*,
-    adminschnittstellen_gremien::*, adminschnittstellen_sitzungen::*,
-    adminschnittstellen_vorgnge::*, authentication::*, authentication_keyadder_schnittstellen::*,
-    autoren_unauthorisiert::*, dokumente_unauthorisiert::*, enumerations_unauthorisiert::*,
-    gremien_unauthorisiert::*, kalender_sitzungen_unauthorisiert::*, sitzungen_unauthorisiert::*,
-    unauthorisiert::*, unauthorisiert_vorgnge::*,
-};
 use utils::{init_tracing, shutdown_signal};
 
 pub type DateTime = chrono::DateTime<chrono::Utc>;
